+++ Vector.h
@@ -6,28 +6,28 @@
 class Vect {
   public:
     // Components
-    float x; float y; float z;
+    double x; double y; double z;
 
     // Constructor
-    Vect(float _x = 0, float _y = 0, float _z = 0);
+    Vect(double _x = 0, double _y = 0, double _z = 0);
 
     // Other methods
-    float dot(const Vect& other) const;
+    double dot(const Vect& other) const;
     Vect cross(const Vect& other) const;
-    float norm() const;
-    float angle(const Vect& other) const;
+    double norm() const;
+    double angle(const Vect& other) const;
     Vect normalize() const;
     String str() const;
 
     // Operator overloads
-    float& operator[](int index);
-    const float& operator[](int index) const;
+    double& operator[](int index);
+    const double& operator[](int index) const;
     Vect operator+(const Vect& other) const;
     Vect operator-(const Vect& other) const;
-    Vect operator*(float other) const;
+    Vect operator*(double other) const;
     Vect& operator+=(const Vect& other);
     Vect& operator-=(const Vect& other);
-    Vect& operator*=(float other);
+    Vect& operator*=(double other);
     void print() const;
 };
